chore: move config definitions into config.rs
